requirements.txt
streamlit
google-generativeai
pymongo
dnspython
